Google_AI_Agent_course/Basic_agents/pyproject.toml
[project]
name = "basic-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "dotenv>=0.9.9",
    "google-adk>=1.19.0",
]
